requirements.txt
Django==3.2.10
django-filter==21.1
django-extensions==3.1.5
django-celery-beat==2.2.1
djangorestframework==3.13.1
drf-yasg==1.20.0
celery==5.2.1
flower==1.0.0
psycopg2-binary==2.9.2
requests==2.26.0
sentry-sdk==1.5.1
gevent==21.1.2
gunicorn

